Changed length to long

+++ Task.java
@@ -2,9 +2,9 @@
 public class Task {
     String name;
     Time dueDate;
-    int length;
+    long length;
 
-    public Task(String name, Time dueDate, int length) {
+    public Task(String name, Time dueDate, long length) {
         this.name = name;
         this.dueDate = dueDate;
         this.length = length;
